:recycle: Move pullRequestReminder to pr_review_reminder.go
because pullRequestReminder no longer depend on http
