+++ main.go
@@ -9,11 +9,9 @@ import (
 	"regexp"
 	"strings"
 
-	"golang.org/x/oauth2"
 	"google.golang.org/appengine"
 	"google.golang.org/appengine/log"
 
-	"github.com/google/go-github/github"
 	"github.com/nlopes/slack"
 	"github.com/nlopes/slack/slackevents"
 )
@@ -139,50 +137,6 @@ func replyToPRReviewReminderMentioned(ctx context.Context, r *http.Request, even
 		reminder.write(b)
 		return b.String()
 	}
-}
-
-func pullRequestReminder(ctx context.Context, team *SlackTeam) (*PRReviewReminder, error) {
-	githubAuthToken, err := GetConfig(ctx, "GITHUB_AUTH_TOKEN")
-	if err != nil {
-		return nil, fmt.Errorf("Failed to get GITHUB_AUTH_TOKEN because of %v", err)
-	}
-	ts := oauth2.StaticTokenSource(
-		&oauth2.Token{AccessToken: githubAuthToken},
-	)
-	tc := oauth2.NewClient(ctx, ts)
-	client := github.NewClient(tc)
-
-	// {"UserLogin": "PR URL"}
-	sum := map[string][]string{}
-	for _, repo := range team.Repositories {
-		userToURLs, err := repo.getUserToReviewUrls(ctx, client)
-		if err != nil {
-			return nil, err
-		}
-		for user, urls := range userToURLs {
-			if sum[user] == nil {
-				sum[user] = []string{}
-			}
-			sum[user] = append(sum[user], urls...)
-		}
-	}
-
-	// https://github.com/nlopes/slack
-	accessToken, err := GetConfig(ctx, "SLACK_OAUTH_ACCESS_TOKEN")
-	if err != nil {
-		return nil, fmt.Errorf("Failed to get SLACK_OAUTH_ACCESS_TOKEN because of %v", err)
-	}
-
-	slack_api := slackApi(ctx, accessToken)
-	userNameToID, err := getUserNameToID(ctx, slack_api)
-	if err != nil {
-		return nil, err
-	}
-
-	return &PRReviewReminder{
-		UserToReviewUrls: sum,
-		UserNameToID:     userNameToID,
-	}, nil
 }
 
 func showPRReviewReminder(w http.ResponseWriter, r *http.Request) {
